Improve the file Movie Facade in order to be more specific. Created a new file called Network Information which have the details for the Query Parameters and the other network related data.
