+++ TheMovieDb/Controller/MovieFacade.swift
@@ -7,48 +7,19 @@
 
 import Foundation
 
-enum MovieListEndpoint {
-    case trending
-    case nowPlaying
-    case popular
-    case topRated
-    case upcoming
-    
-    fileprivate var path: String {
-        switch self {
-        case .trending: return "trending/movie/week"
-        case .nowPlaying: return "movie/now_playing"
-        case .popular: return "movie/popular"
-        case .topRated: return "movie/top_rated"
-        case .upcoming: return "movie/upcoming"
-        }
-    }
-}
-enum HTTPMethod: String {
-    case get, post, put, patch, delete, options, head
-}
-
-enum MovieError: Error {
-    case invalidUrl
-    case wrongResponse
-    case unknownError(error: Error)
-}
-
-protocol MovieService {
-    static func getMovies(endpoint: MovieListEndpoint, returnMovies: @escaping (Result<MovieResponse, MovieError>) -> Void)
-}
-
-struct MovieFacade: MovieService {
+struct MovieFacade: MovieService, QueryParams {
     
     static let baseURL = URL(string:"https://api.themoviedb.org/3/")!
-    static let apiKey = "f6cd5c1a9e6c6b965fdcab0fa6ddd38a"
+    static var apiKey = "f6cd5c1a9e6c6b965fdcab0fa6ddd38a"
+    static var language = "en-US"
     
     static func getMovies(endpoint: MovieListEndpoint, returnMovies: @escaping (Result<MovieResponse, MovieError>) -> Void) {
 
         var components = URLComponents()
         components.path = endpoint.path
         components.queryItems = [
-            .init(name: "api_key", value: apiKey)
+            .init(name: "api_key", value: apiKey),
+            .init(name: "language", value: language)
         ]
         
         guard let url = components.url(relativeTo: baseURL) else {
